v0.14.0 General Restructurization & AR
The project's folder structure and assembly references were heavily changed to only what is necessary.

A new type of burst was added: An automatic burst. This one works while holding the button, contrary to the manual variant.

A new weapon was added: A semi-automatic Assault Rifle. It shoots a burst of 3 bullets.

Removed or changed unnecesary parts of the code, such as minor errors in PlayerCameraController.cs, removed the GetCurrentWeapon method in WeaponManagerScript.cs, and some others.