v0.15.0 Gun Ammo & Pickup
Weapons now hold a limited amount of ammo, and need to reload after firing all the bullets in a magazine.

An entity with a weapon manager can walk over a pickup point and their total ammo will increment by the set amount, but this amount can never exceed the maximum amount of ammo set for each weapon.

The pickup point has a sprite that rotates to look at the camera through Billboarding.