v0.16.0 Ammo UI Bars
There are now ammo bars in the UI that track the ammo of the current weapon. Bars don't appear until the first time they are needed, and don't disappear. This is intentional.

This was done by using events in several different scripts.

Why can't we just have the bar script read what is needed directly from the weapon script without also reading from the Weapon Manager?
Because we need to keep track of the weapon in case it has changed. Otherwise, we would only update one weapon and leave the rest un-updated.

This still needs further optimization, but it works as intended for now.

We now store all the Weapon Scripts in the Weapon Manager class in order to reduce the amount of GetComponent calls in the long run.

An option was added for each weapon to have infinite ammo, although it would be more precise to say that it doesn't consume ammo, since it still needs at least 1 bullet to shoot.