v0.16.1 Ammo UI Bars Optimization
Through changing some parts of the code, we now do basically the same thing with less code.

Unnecessary events were removed, and we only use a single event now to handle all updates.

This event, originated in the WeaponScript, is listened by the WeaponManagerScript, which in turn has an event that the WeaponReloadBarScript listens to.