v0.16.2 Automatic Reload
Weapons can now reload their ammo automatically, decided based on what the proggramer decides at compile time.

Yes, this could be another variable for the weapon, but I believe that automatic reload should remain consistent, meanning that if one weapon has it, every one should also have it and vice versa.

Whether to enable the feature or not is decided by the proggrammer.

Also updated the player and marker prefabs.